Adding a warning also for the depth dimension (#72)
We detected that there was no warning when the depth dimension was being set out of bounds... we corrected the code, so that it raises a warning when the requested interval is out. 
Careful, this is a breaking change as it can change the output. It can also change the behaviour if --subset-method is set to strict.